clean the directory
