--- benchmarks/backends/ProjectQ/benchmark.py
@@ -1,1444 +1,0 @@
-import projectq
-import math
-import time
-import sys
-from projectq.backends import Simulator
-from projectq.ops import H, Rz, Ry, CNOT, Measure
-eng = projectq.MainEngine(Simulator(gate_fusion=False))
-start = time.time()
-q0 = eng.allocate_qureg(10)
-c0 = [0] * 10
-#U(1.14215051891182,-1.11411950570152,1.79622558568019) q0[2];
-Rz(1.14215051891182) | q0[2];
-Ry(-1.11411950570152) | q0[2];
-Rz(1.79622558568019) | q0[2];
-#U(1.74856095647920,-1.69942443278862,-2.07613929130468) q0[7];
-Rz(1.74856095647920) | q0[7];
-Ry(-1.69942443278862) | q0[7];
-Rz(-2.07613929130468) | q0[7];
-#CX q0[7],q0[2];
-CNOT | (q0[7], q0[2])
-#U(0,0,0.861907904839883) q0[2];
-Rz(0.861907904839883) | q0[2];
-#U(-0.428088935792326,0.0,0.0) q0[7];
-Rz(-0.428088935792326) | q0[7];
-#CX q0[2],q0[7];
-CNOT | (q0[2], q0[7])
-#U(3.04395932545489,0.0,0.0) q0[7];
-Rz(3.04395932545489) | q0[7];
-#CX q0[7],q0[2];
-CNOT | (q0[7], q0[2])
-#U(1.04399839462527,2.52458764482196,-3.05688304492963) q0[2];
-Rz(1.04399839462527) | q0[2];
-Ry(2.52458764482196) | q0[2];
-Rz(-3.05688304492963) | q0[2];
-#U(1.23664101537033,-1.52641135384280,-1.36928143652801) q0[7];
-Rz(1.23664101537033) | q0[7];
-Ry(-1.52641135384280) | q0[7];
-Rz(-1.36928143652801) | q0[7];
-#U(2.24812222386752,0.164819681570933,0.344112601489393) q0[9];
-Rz(2.24812222386752) | q0[9];
-Ry(0.164819681570933) | q0[9];
-Rz(0.344112601489393) | q0[9];
-#U(1.12444268366144,-3.45724569730781,-1.15042701939009) q0[8];
-Rz(1.12444268366144) | q0[8];
-Ry(-3.45724569730781) | q0[8];
-Rz(-1.15042701939009) | q0[8];
-#CX q0[8],q0[9];
-CNOT | (q0[8], q0[9])
-#U(0,0,-0.602886496766072) q0[9];
-Rz(-0.602886496766072) | q0[9];
-#U(1.43092668677630,0.0,0.0) q0[8];
-Rz(1.43092668677630) | q0[8];
-#CX q0[9],q0[8];
-CNOT | (q0[9], q0[8])
-#U(3.65417973085514,0.0,0.0) q0[8];
-Rz(3.65417973085514) | q0[8];
-#CX q0[8],q0[9];
-CNOT | (q0[8], q0[9])
-#U(2.47216960935367,2.22695758203971,0.383106789768156) q0[9];
-Rz(2.47216960935367) | q0[9];
-Ry(2.22695758203971) | q0[9];
-Rz(0.383106789768156) | q0[9];
-#U(2.50251653461346,-2.50140545788134,2.02500407586615) q0[8];
-Rz(2.50251653461346) | q0[8];
-Ry(-2.50140545788134) | q0[8];
-Rz(2.02500407586615) | q0[8];
-#U(2.03020528191708,2.59862476942509,-0.868062865180180) q0[0];
-Rz(2.03020528191708) | q0[0];
-Ry(2.59862476942509) | q0[0];
-Rz(-0.868062865180180) | q0[0];
-#U(2.32132812684660,1.04507296075213,-2.70630669882031) q0[3];
-Rz(2.32132812684660) | q0[3];
-Ry(1.04507296075213) | q0[3];
-Rz(-2.70630669882031) | q0[3];
-#CX q0[3],q0[0];
-CNOT | (q0[3], q0[0])
-#U(0,0,-0.940787453803027) q0[0];
-Rz(-0.940787453803027) | q0[0];
-#U(0.452138136122875,0.0,0.0) q0[3];
-Rz(0.452138136122875) | q0[3];
-#CX q0[0],q0[3];
-CNOT | (q0[0], q0[3])
-#U(3.81016152921993,0.0,0.0) q0[3];
-Rz(3.81016152921993) | q0[3];
-#CX q0[3],q0[0];
-CNOT | (q0[3], q0[0])
-#U(2.06214718164040,-2.97817905762658,0.0717362571192934) q0[0];
-Rz(2.06214718164040) | q0[0];
-Ry(-2.97817905762658) | q0[0];
-Rz(0.0717362571192934) | q0[0];
-#U(2.04848000513242,-0.326173544802899,-2.87162656408750) q0[3];
-Rz(2.04848000513242) | q0[3];
-Ry(-0.326173544802899) | q0[3];
-Rz(-2.87162656408750) | q0[3];
-#U(1.69563349670311,-2.03269798500633,4.17588143457218) q0[1];
-Rz(1.69563349670311) | q0[1];
-Ry(-2.03269798500633) | q0[1];
-Rz(4.17588143457218) | q0[1];
-#U(2.15672592016448,1.42212855034625,-1.89225089382840) q0[6];
-Rz(2.15672592016448) | q0[6];
-Ry(1.42212855034625) | q0[6];
-Rz(-1.89225089382840) | q0[6];
-#CX q0[6],q0[1];
-CNOT | (q0[6], q0[1])
-#U(0,0,3.20394217614098) q0[1];
-Rz(3.20394217614098) | q0[1];
-#U(-1.94728979824532,0.0,0.0) q0[6];
-Rz(-1.94728979824532) | q0[6];
-#CX q0[1],q0[6];
-CNOT | (q0[1], q0[6])
-#U(1.18027358199685,0.0,0.0) q0[6];
-Rz(1.18027358199685) | q0[6];
-#CX q0[6],q0[1];
-CNOT | (q0[6], q0[1])
-#U(1.26528922827283,-0.787275539595522,0.674357049932829) q0[1];
-Rz(1.26528922827283) | q0[1];
-Ry(-0.787275539595522) | q0[1];
-Rz(0.674357049932829) | q0[1];
-#U(1.10942048619383,1.08347971734110,4.24772936273010) q0[6];
-Rz(1.10942048619383) | q0[6];
-Ry(1.08347971734110) | q0[6];
-Rz(4.24772936273010) | q0[6];
-#U(1.52417602137264,-1.82268169237610,-0.397038647053302) q0[4];
-Rz(1.52417602137264) | q0[4];
-Ry(-1.82268169237610) | q0[4];
-Rz(-0.397038647053302) | q0[4];
-#U(1.41480177969943,-3.25712577023569,-1.02814085787792) q0[5];
-Rz(1.41480177969943) | q0[5];
-Ry(-3.25712577023569) | q0[5];
-Rz(-1.02814085787792) | q0[5];
-#CX q0[5],q0[4];
-CNOT | (q0[5], q0[4])
-#U(0,0,3.61866699761798) q0[4];
-Rz(3.61866699761798) | q0[4];
-#U(-3.44378696290912,0.0,0.0) q0[5];
-Rz(-3.44378696290912) | q0[5];
-#CX q0[4],q0[5];
-CNOT | (q0[4], q0[5])
-#U(-1.33699053056508,0.0,0.0) q0[5];
-Rz(-1.33699053056508) | q0[5];
-#CX q0[5],q0[4];
-CNOT | (q0[5], q0[4])
-#U(2.49140177690712,1.59335454632054,-2.61602079225387) q0[4];
-Rz(2.49140177690712) | q0[4];
-Ry(1.59335454632054) | q0[4];
-Rz(-2.61602079225387) | q0[4];
-#U(2.48539448715742,2.37978821853565,1.48424714110929) q0[5];
-Rz(2.48539448715742) | q0[5];
-Ry(2.37978821853565) | q0[5];
-Rz(1.48424714110929) | q0[5];
-#U(1.96987818120257,0.747401677098067,0.704996001115372) q0[0];
-Rz(1.96987818120257) | q0[0];
-Ry(0.747401677098067) | q0[0];
-Rz(0.704996001115372) | q0[0];
-#U(1.33043369127056,-2.05258846043485,-2.19120028144956) q0[2];
-Rz(1.33043369127056) | q0[2];
-Ry(-2.05258846043485) | q0[2];
-Rz(-2.19120028144956) | q0[2];
-#CX q0[2],q0[0];
-CNOT | (q0[2], q0[0])
-#U(0,0,2.98354841656014) q0[0];
-Rz(2.98354841656014) | q0[0];
-#U(-1.28579017816580,0.0,0.0) q0[2];
-Rz(-1.28579017816580) | q0[2];
-#CX q0[0],q0[2];
-CNOT | (q0[0], q0[2])
-#U(2.47358054449452,0.0,0.0) q0[2];
-Rz(2.47358054449452) | q0[2];
-#CX q0[2],q0[0];
-CNOT | (q0[2], q0[0])
-#U(2.67477182638161,-0.616717709940549,1.69524652589156) q0[0];
-Rz(2.67477182638161) | q0[0];
-Ry(-0.616717709940549) | q0[0];
-Rz(1.69524652589156) | q0[0];
-#U(1.65378501252131,-1.70723595214641,4.47665400056241) q0[2];
-Rz(1.65378501252131) | q0[2];
-Ry(-1.70723595214641) | q0[2];
-Rz(4.47665400056241) | q0[2];
-#U(1.30805580368873,2.26717281757197,0.304560220245233) q0[7];
-Rz(1.30805580368873) | q0[7];
-Ry(2.26717281757197) | q0[7];
-Rz(0.304560220245233) | q0[7];
-#U(0.513473976439764,0.910162222977109,-4.13155316380043) q0[8];
-Rz(0.513473976439764) | q0[8];
-Ry(0.910162222977109) | q0[8];
-Rz(-4.13155316380043) | q0[8];
-#CX q0[8],q0[7];
-CNOT | (q0[8], q0[7])
-#U(0,0,3.15980569971129) q0[7];
-Rz(3.15980569971129) | q0[7];
-#U(-0.532128036916324,0.0,0.0) q0[8];
-Rz(-0.532128036916324) | q0[8];
-#CX q0[7],q0[8];
-CNOT | (q0[7], q0[8])
-#U(1.97324987644596,0.0,0.0) q0[8];
-Rz(1.97324987644596) | q0[8];
-#CX q0[8],q0[7];
-CNOT | (q0[8], q0[7])
-#U(1.26173622403787,-0.368802695505360,-2.33664647768154) q0[7];
-Rz(1.26173622403787) | q0[7];
-Ry(-0.368802695505360) | q0[7];
-Rz(-2.33664647768154) | q0[7];
-#U(2.56622331210234,-2.32238566866662,3.88837680590712) q0[8];
-Rz(2.56622331210234) | q0[8];
-Ry(-2.32238566866662) | q0[8];
-Rz(3.88837680590712) | q0[8];
-#U(1.21624104997990,-0.855390450092245,1.43059743578507) q0[6];
-Rz(1.21624104997990) | q0[6];
-Ry(-0.855390450092245) | q0[6];
-Rz(1.43059743578507) | q0[6];
-#U(1.89260290392798,-1.66016343496994,-2.28800135960327) q0[5];
-Rz(1.89260290392798) | q0[5];
-Ry(-1.66016343496994) | q0[5];
-Rz(-2.28800135960327) | q0[5];
-#CX q0[5],q0[6];
-CNOT | (q0[5], q0[6])
-#U(0,0,1.07448852068098) q0[6];
-Rz(1.07448852068098) | q0[6];
-#U(-1.55843359831872,0.0,0.0) q0[5];
-Rz(-1.55843359831872) | q0[5];
-#CX q0[6],q0[5];
-CNOT | (q0[6], q0[5])
-#U(2.83209070667356,0.0,0.0) q0[5];
-Rz(2.83209070667356) | q0[5];
-#CX q0[5],q0[6];
-CNOT | (q0[5], q0[6])
-#U(0.913634810756815,2.43311718342605,-0.261257971109959) q0[6];
-Rz(0.913634810756815) | q0[6];
-Ry(2.43311718342605) | q0[6];
-Rz(-0.261257971109959) | q0[6];
-#U(2.09297123496674,-0.904716446019859,-4.30707950845936) q0[5];
-Rz(2.09297123496674) | q0[5];
-Ry(-0.904716446019859) | q0[5];
-Rz(-4.30707950845936) | q0[5];
-#U(1.46482045649967,-0.329672629956639,0.451175355985203) q0[1];
-Rz(1.46482045649967) | q0[1];
-Ry(-0.329672629956639) | q0[1];
-Rz(0.451175355985203) | q0[1];
-#U(2.02596449647726,-3.59551200279199,-0.415795885523276) q0[3];
-Rz(2.02596449647726) | q0[3];
-Ry(-3.59551200279199) | q0[3];
-Rz(-0.415795885523276) | q0[3];
-#CX q0[3],q0[1];
-CNOT | (q0[3], q0[1])
-#U(0,0,2.39823965904581) q0[1];
-Rz(2.39823965904581) | q0[1];
-#U(-1.68614605540418,0.0,0.0) q0[3];
-Rz(-1.68614605540418) | q0[3];
-#CX q0[1],q0[3];
-CNOT | (q0[1], q0[3])
-#U(1.12019636302332,0.0,0.0) q0[3];
-Rz(1.12019636302332) | q0[3];
-#CX q0[3],q0[1];
-CNOT | (q0[3], q0[1])
-#U(3.06912917524574,3.53705489883201,-1.52636284946318) q0[1];
-Rz(3.06912917524574) | q0[1];
-Ry(3.53705489883201) | q0[1];
-Rz(-1.52636284946318) | q0[1];
-#U(0.334805085203238,0.713656027655973,-3.12743406555175) q0[3];
-Rz(0.334805085203238) | q0[3];
-Ry(0.713656027655973) | q0[3];
-Rz(-3.12743406555175) | q0[3];
-#U(1.86240531261674,1.29861689798061,-3.02412454564690) q0[4];
-Rz(1.86240531261674) | q0[4];
-Ry(1.29861689798061) | q0[4];
-Rz(-3.02412454564690) | q0[4];
-#U(0.825087365512213,-2.40387515598303,2.34523979737775) q0[9];
-Rz(0.825087365512213) | q0[9];
-Ry(-2.40387515598303) | q0[9];
-Rz(2.34523979737775) | q0[9];
-#CX q0[9],q0[4];
-CNOT | (q0[9], q0[4])
-#U(0,0,-0.271427281543746) q0[4];
-Rz(-0.271427281543746) | q0[4];
-#U(-1.60918560682224,0.0,0.0) q0[9];
-Rz(-1.60918560682224) | q0[9];
-#CX q0[4],q0[9];
-CNOT | (q0[4], q0[9])
-#U(0.782531495105594,0.0,0.0) q0[9];
-Rz(0.782531495105594) | q0[9];
-#CX q0[9],q0[4];
-CNOT | (q0[9], q0[4])
-#U(0.715189642190039,1.99660355652784,-2.67159395656149) q0[4];
-Rz(0.715189642190039) | q0[4];
-Ry(1.99660355652784) | q0[4];
-Rz(-2.67159395656149) | q0[4];
-#U(2.02236904492581,1.16814483944193,-2.29201302661145) q0[9];
-Rz(2.02236904492581) | q0[9];
-Ry(1.16814483944193) | q0[9];
-Rz(-2.29201302661145) | q0[9];
-#U(2.28207870743267,-0.983636995518970,0.183086569837585) q0[1];
-Rz(2.28207870743267) | q0[1];
-Ry(-0.983636995518970) | q0[1];
-Rz(0.183086569837585) | q0[1];
-#U(2.04217371543296,-2.68081876238919,-0.527240995389771) q0[0];
-Rz(2.04217371543296) | q0[0];
-Ry(-2.68081876238919) | q0[0];
-Rz(-0.527240995389771) | q0[0];
-#CX q0[0],q0[1];
-CNOT | (q0[0], q0[1])
-#U(0,0,-0.591110353301228) q0[1];
-Rz(-0.591110353301228) | q0[1];
-#U(-1.67519341634164,0.0,0.0) q0[0];
-Rz(-1.67519341634164) | q0[0];
-#CX q0[1],q0[0];
-CNOT | (q0[1], q0[0])
-#U(1.05110298017014,0.0,0.0) q0[0];
-Rz(1.05110298017014) | q0[0];
-#CX q0[0],q0[1];
-CNOT | (q0[0], q0[1])
-#U(1.13380880241054,-1.62718349944808,-0.803121984177828) q0[1];
-Rz(1.13380880241054) | q0[1];
-Ry(-1.62718349944808) | q0[1];
-Rz(-0.803121984177828) | q0[1];
-#U(0.662410345089536,1.45258658138185,2.28574298287533) q0[0];
-Rz(0.662410345089536) | q0[0];
-Ry(1.45258658138185) | q0[0];
-Rz(2.28574298287533) | q0[0];
-#U(2.76023830319668,2.95083953512843,-1.27838294143947) q0[6];
-Rz(2.76023830319668) | q0[6];
-Ry(2.95083953512843) | q0[6];
-Rz(-1.27838294143947) | q0[6];
-#U(1.89248677820431,-0.563312698725226,-5.39933491773262) q0[2];
-Rz(1.89248677820431) | q0[2];
-Ry(-0.563312698725226) | q0[2];
-Rz(-5.39933491773262) | q0[2];
-#CX q0[2],q0[6];
-CNOT | (q0[2], q0[6])
-#U(0,0,0.789860299901999) q0[6];
-Rz(0.789860299901999) | q0[6];
-#U(-1.33942885442794,0.0,0.0) q0[2];
-Rz(-1.33942885442794) | q0[2];
-#CX q0[6],q0[2];
-CNOT | (q0[6], q0[2])
-#U(-0.116680370130935,0.0,0.0) q0[2];
-Rz(-0.116680370130935) | q0[2];
-#CX q0[2],q0[6];
-CNOT | (q0[2], q0[6])
-#U(2.19481104574862,-0.672348668018873,1.87810698390264) q0[6];
-Rz(2.19481104574862) | q0[6];
-Ry(-0.672348668018873) | q0[6];
-Rz(1.87810698390264) | q0[6];
-#U(1.49276670608634,4.38704347212247,-1.58612693480310) q0[2];
-Rz(1.49276670608634) | q0[2];
-Ry(4.38704347212247) | q0[2];
-Rz(-1.58612693480310) | q0[2];
-#U(1.16006893932404,3.20811412686438,-2.04329743665573) q0[4];
-Rz(1.16006893932404) | q0[4];
-Ry(3.20811412686438) | q0[4];
-Rz(-2.04329743665573) | q0[4];
-#U(1.59804490504680,0.996650448594488,-0.680131163823329) q0[3];
-Rz(1.59804490504680) | q0[3];
-Ry(0.996650448594488) | q0[3];
-Rz(-0.680131163823329) | q0[3];
-#CX q0[3],q0[4];
-CNOT | (q0[3], q0[4])
-#U(0,0,1.41171030510075) q0[4];
-Rz(1.41171030510075) | q0[4];
-#U(-0.473337386122504,0.0,0.0) q0[3];
-Rz(-0.473337386122504) | q0[3];
-#CX q0[4],q0[3];
-CNOT | (q0[4], q0[3])
-#U(-0.305791249353728,0.0,0.0) q0[3];
-Rz(-0.305791249353728) | q0[3];
-#CX q0[3],q0[4];
-CNOT | (q0[3], q0[4])
-#U(0.857354960076594,-0.850543059384741,1.54528206261559) q0[4];
-Rz(0.857354960076594) | q0[4];
-Ry(-0.850543059384741) | q0[4];
-Rz(1.54528206261559) | q0[4];
-#U(2.21422611666955,4.11432477937595,-0.898528363036821) q0[3];
-Rz(2.21422611666955) | q0[3];
-Ry(4.11432477937595) | q0[3];
-Rz(-0.898528363036821) | q0[3];
-#U(1.31853259891572,1.50298910777540,-2.64660171296098) q0[5];
-Rz(1.31853259891572) | q0[5];
-Ry(1.50298910777540) | q0[5];
-Rz(-2.64660171296098) | q0[5];
-#U(1.52753668855680,-3.76374234358738,2.37075853267450) q0[7];
-Rz(1.52753668855680) | q0[7];
-Ry(-3.76374234358738) | q0[7];
-Rz(2.37075853267450) | q0[7];
-#CX q0[7],q0[5];
-CNOT | (q0[7], q0[5])
-#U(0,0,3.18262336776567) q0[5];
-Rz(3.18262336776567) | q0[5];
-#U(-2.33766634664937,0.0,0.0) q0[7];
-Rz(-2.33766634664937) | q0[7];
-#CX q0[5],q0[7];
-CNOT | (q0[5], q0[7])
-#U(1.20027560182840,0.0,0.0) q0[7];
-Rz(1.20027560182840) | q0[7];
-#CX q0[7],q0[5];
-CNOT | (q0[7], q0[5])
-#U(1.78527770562790,1.79094858777953,-3.59026081693414) q0[5];
-Rz(1.78527770562790) | q0[5];
-Ry(1.79094858777953) | q0[5];
-Rz(-3.59026081693414) | q0[5];
-#U(0.618189008543180,4.13382440828138,2.04238353222969) q0[7];
-Rz(0.618189008543180) | q0[7];
-Ry(4.13382440828138) | q0[7];
-Rz(2.04238353222969) | q0[7];
-#U(0.724009318650646,-0.648328044764882,1.81258285221843) q0[9];
-Rz(0.724009318650646) | q0[9];
-Ry(-0.648328044764882) | q0[9];
-Rz(1.81258285221843) | q0[9];
-#U(1.03244916404794,-2.08516478333061,-0.517916366394120) q0[8];
-Rz(1.03244916404794) | q0[8];
-Ry(-2.08516478333061) | q0[8];
-Rz(-0.517916366394120) | q0[8];
-#CX q0[8],q0[9];
-CNOT | (q0[8], q0[9])
-#U(0,0,-0.0467489466862110) q0[9];
-Rz(-0.0467489466862110) | q0[9];
-#U(-0.692237346093271,0.0,0.0) q0[8];
-Rz(-0.692237346093271) | q0[8];
-#CX q0[9],q0[8];
-CNOT | (q0[9], q0[8])
-#U(2.19810050176112,0.0,0.0) q0[8];
-Rz(2.19810050176112) | q0[8];
-#CX q0[8],q0[9];
-CNOT | (q0[8], q0[9])
-#U(2.23564793162950,3.07400854925085,-0.758846920715642) q0[9];
-Rz(2.23564793162950) | q0[9];
-Ry(3.07400854925085) | q0[9];
-Rz(-0.758846920715642) | q0[9];
-#U(1.90982332151684,-1.34865350506798,-4.01567161189411) q0[8];
-Rz(1.90982332151684) | q0[8];
-Ry(-1.34865350506798) | q0[8];
-Rz(-4.01567161189411) | q0[8];
-#U(1.86438896146350,0.720642635814958,-0.959553038653771) q0[4];
-Rz(1.86438896146350) | q0[4];
-Ry(0.720642635814958) | q0[4];
-Rz(-0.959553038653771) | q0[4];
-#U(0.646791620197337,0.924687241170619,-4.03970909287212) q0[3];
-Rz(0.646791620197337) | q0[3];
-Ry(0.924687241170619) | q0[3];
-Rz(-4.03970909287212) | q0[3];
-#CX q0[3],q0[4];
-CNOT | (q0[3], q0[4])
-#U(0,0,-0.231373159097814) q0[4];
-Rz(-0.231373159097814) | q0[4];
-#U(1.28073002593365,0.0,0.0) q0[3];
-Rz(1.28073002593365) | q0[3];
-#CX q0[4],q0[3];
-CNOT | (q0[4], q0[3])
-#U(3.65481080875041,0.0,0.0) q0[3];
-Rz(3.65481080875041) | q0[3];
-#CX q0[3],q0[4];
-CNOT | (q0[3], q0[4])
-#U(2.33327938240083,3.13886104310805,-0.393990739817776) q0[4];
-Rz(2.33327938240083) | q0[4];
-Ry(3.13886104310805) | q0[4];
-Rz(-0.393990739817776) | q0[4];
-#U(0.958982320896732,-1.86553995730297,0.259620690790581) q0[3];
-Rz(0.958982320896732) | q0[3];
-Ry(-1.86553995730297) | q0[3];
-Rz(0.259620690790581) | q0[3];
-#U(2.66539491185486,1.16217464026759,-2.94248105549771) q0[6];
-Rz(2.66539491185486) | q0[6];
-Ry(1.16217464026759) | q0[6];
-Rz(-2.94248105549771) | q0[6];
-#U(1.98700026493144,-2.71773734908148,2.66341971448259) q0[2];
-Rz(1.98700026493144) | q0[2];
-Ry(-2.71773734908148) | q0[2];
-Rz(2.66341971448259) | q0[2];
-#CX q0[2],q0[6];
-CNOT | (q0[2], q0[6])
-#U(0,0,1.10843697099204) q0[6];
-Rz(1.10843697099204) | q0[6];
-#U(-0.352507041584004,0.0,0.0) q0[2];
-Rz(-0.352507041584004) | q0[2];
-#CX q0[6],q0[2];
-CNOT | (q0[6], q0[2])
-#U(2.59676135890237,0.0,0.0) q0[2];
-Rz(2.59676135890237) | q0[2];
-#CX q0[2],q0[6];
-CNOT | (q0[2], q0[6])
-#U(2.30297132778683,-0.419494636687877,1.50439628657172) q0[6];
-Rz(2.30297132778683) | q0[6];
-Ry(-0.419494636687877) | q0[6];
-Rz(1.50439628657172) | q0[6];
-#U(1.88751520354741,-5.47477162629485,-0.530474564671800) q0[2];
-Rz(1.88751520354741) | q0[2];
-Ry(-5.47477162629485) | q0[2];
-Rz(-0.530474564671800) | q0[2];
-#U(2.86165608056341,3.98436512503058,-2.15070694233033) q0[5];
-Rz(2.86165608056341) | q0[5];
-Ry(3.98436512503058) | q0[5];
-Rz(-2.15070694233033) | q0[5];
-#U(1.58033111643771,-0.251200035747908,2.59567943850498) q0[7];
-Rz(1.58033111643771) | q0[7];
-Ry(-0.251200035747908) | q0[7];
-Rz(2.59567943850498) | q0[7];
-#CX q0[7],q0[5];
-CNOT | (q0[7], q0[5])
-#U(0,0,0.608325714280855) q0[5];
-Rz(0.608325714280855) | q0[5];
-#U(-1.41253614693639,0.0,0.0) q0[7];
-Rz(-1.41253614693639) | q0[7];
-#CX q0[5],q0[7];
-CNOT | (q0[5], q0[7])
-#U(2.85545722662107,0.0,0.0) q0[7];
-Rz(2.85545722662107) | q0[7];
-#CX q0[7],q0[5];
-CNOT | (q0[7], q0[5])
-#U(1.78409276095002,-1.22919721488071,1.11279636916952) q0[5];
-Rz(1.78409276095002) | q0[5];
-Ry(-1.22919721488071) | q0[5];
-Rz(1.11279636916952) | q0[5];
-#U(0.313683229213436,-0.611575221204724,4.15127791573039) q0[7];
-Rz(0.313683229213436) | q0[7];
-Ry(-0.611575221204724) | q0[7];
-Rz(4.15127791573039) | q0[7];
-#U(0.823589397226933,-1.00089510205415,0.643337578418273) q0[8];
-Rz(0.823589397226933) | q0[8];
-Ry(-1.00089510205415) | q0[8];
-Rz(0.643337578418273) | q0[8];
-#U(0.740112149390534,-2.02814175218469,-0.972490309311748) q0[9];
-Rz(0.740112149390534) | q0[9];
-Ry(-2.02814175218469) | q0[9];
-Rz(-0.972490309311748) | q0[9];
-#CX q0[9],q0[8];
-CNOT | (q0[9], q0[8])
-#U(0,0,2.32487109247705) q0[8];
-Rz(2.32487109247705) | q0[8];
-#U(-0.0468199409109238,0.0,0.0) q0[9];
-Rz(-0.0468199409109238) | q0[9];
-#CX q0[8],q0[9];
-CNOT | (q0[8], q0[9])
-#U(1.41644878892005,0.0,0.0) q0[9];
-Rz(1.41644878892005) | q0[9];
-#CX q0[9],q0[8];
-CNOT | (q0[9], q0[8])
-#U(0.809496683520842,0.657623240321232,1.01309988015780) q0[8];
-Rz(0.809496683520842) | q0[8];
-Ry(0.657623240321232) | q0[8];
-Rz(1.01309988015780) | q0[8];
-#U(1.42500250245906,-4.11797852109158,0.222480277485207) q0[9];
-Rz(1.42500250245906) | q0[9];
-Ry(-4.11797852109158) | q0[9];
-Rz(0.222480277485207) | q0[9];
-#U(2.07361985931839,0.461147477923885,1.45611118617045) q0[1];
-Rz(2.07361985931839) | q0[1];
-Ry(0.461147477923885) | q0[1];
-Rz(1.45611118617045) | q0[1];
-#U(1.34457240597485,-2.61401452265593,-1.72478088140032) q0[0];
-Rz(1.34457240597485) | q0[0];
-Ry(-2.61401452265593) | q0[0];
-Rz(-1.72478088140032) | q0[0];
-#CX q0[0],q0[1];
-CNOT | (q0[0], q0[1])
-#U(0,0,1.46527087800235) q0[1];
-Rz(1.46527087800235) | q0[1];
-#U(-0.150559053513166,0.0,0.0) q0[0];
-Rz(-0.150559053513166) | q0[0];
-#CX q0[1],q0[0];
-CNOT | (q0[1], q0[0])
-#U(2.03743357557980,0.0,0.0) q0[0];
-Rz(2.03743357557980) | q0[0];
-#CX q0[0],q0[1];
-CNOT | (q0[0], q0[1])
-#U(2.19762103178133,0.324461760385907,-3.20241359681867) q0[1];
-Rz(2.19762103178133) | q0[1];
-Ry(0.324461760385907) | q0[1];
-Rz(-3.20241359681867) | q0[1];
-#U(1.64569941525527,5.65202677976156,-0.208968745485691) q0[0];
-Rz(1.64569941525527) | q0[0];
-Ry(5.65202677976156) | q0[0];
-Rz(-0.208968745485691) | q0[0];
-#U(0.631079662861870,-2.71829026971893,-0.0658517672705348) q0[6];
-Rz(0.631079662861870) | q0[6];
-Ry(-2.71829026971893) | q0[6];
-Rz(-0.0658517672705348) | q0[6];
-#U(1.54111947191762,-3.18671165433959,-0.00664180358196198) q0[4];
-Rz(1.54111947191762) | q0[4];
-Ry(-3.18671165433959) | q0[4];
-Rz(-0.00664180358196198) | q0[4];
-#CX q0[4],q0[6];
-CNOT | (q0[4], q0[6])
-#U(0,0,-0.742714502507306) q0[6];
-Rz(-0.742714502507306) | q0[6];
-#U(-1.66415725146374,0.0,0.0) q0[4];
-Rz(-1.66415725146374) | q0[4];
-#CX q0[6],q0[4];
-CNOT | (q0[6], q0[4])
-#U(1.18928448375649,0.0,0.0) q0[4];
-Rz(1.18928448375649) | q0[4];
-#CX q0[4],q0[6];
-CNOT | (q0[4], q0[6])
-#U(1.60338378645355,1.57268955457366,-3.74595376815610) q0[6];
-Rz(1.60338378645355) | q0[6];
-Ry(1.57268955457366) | q0[6];
-Rz(-3.74595376815610) | q0[6];
-#U(1.11641349723153,-3.62753708369334,-0.552335253577286) q0[4];
-Rz(1.11641349723153) | q0[4];
-Ry(-3.62753708369334) | q0[4];
-Rz(-0.552335253577286) | q0[4];
-#U(2.32953318879111,1.26723592760012,-2.97546160357209) q0[8];
-Rz(2.32953318879111) | q0[8];
-Ry(1.26723592760012) | q0[8];
-Rz(-2.97546160357209) | q0[8];
-#U(1.98501482151612,-2.34442810792294,2.50270593423907) q0[1];
-Rz(1.98501482151612) | q0[1];
-Ry(-2.34442810792294) | q0[1];
-Rz(2.50270593423907) | q0[1];
-#CX q0[1],q0[8];
-CNOT | (q0[1], q0[8])
-#U(0,0,-0.795088025984119) q0[8];
-Rz(-0.795088025984119) | q0[8];
-#U(-0.0183904149418799,0.0,0.0) q0[1];
-Rz(-0.0183904149418799) | q0[1];
-#CX q0[8],q0[1];
-CNOT | (q0[8], q0[1])
-#U(3.67917914388315,0.0,0.0) q0[1];
-Rz(3.67917914388315) | q0[1];
-#CX q0[1],q0[8];
-CNOT | (q0[1], q0[8])
-#U(0.234892807255841,1.55634706392549,-3.01173910768412) q0[8];
-Rz(0.234892807255841) | q0[8];
-Ry(1.55634706392549) | q0[8];
-Rz(-3.01173910768412) | q0[8];
-#U(2.13095717390631,4.50022429329697,1.50933728606963) q0[1];
-Rz(2.13095717390631) | q0[1];
-Ry(4.50022429329697) | q0[1];
-Rz(1.50933728606963) | q0[1];
-#U(2.86353119325408,-0.231657016405759,0.271714922609275) q0[9];
-Rz(2.86353119325408) | q0[9];
-Ry(-0.231657016405759) | q0[9];
-Rz(0.271714922609275) | q0[9];
-#U(1.09537043062845,-3.65264368705059,-1.53559777199950) q0[2];
-Rz(1.09537043062845) | q0[2];
-Ry(-3.65264368705059) | q0[2];
-Rz(-1.53559777199950) | q0[2];
-#CX q0[2],q0[9];
-CNOT | (q0[2], q0[9])
-#U(0,0,3.36178266384714) q0[9];
-Rz(3.36178266384714) | q0[9];
-#U(-4.18332147802948,0.0,0.0) q0[2];
-Rz(-4.18332147802948) | q0[2];
-#CX q0[9],q0[2];
-CNOT | (q0[9], q0[2])
-#U(-0.173456430468486,0.0,0.0) q0[2];
-Rz(-0.173456430468486) | q0[2];
-#CX q0[2],q0[9];
-CNOT | (q0[2], q0[9])
-#U(2.70887832601418,0.937497851959002,0.497002943175428) q0[9];
-Rz(2.70887832601418) | q0[9];
-Ry(0.937497851959002) | q0[9];
-Rz(0.497002943175428) | q0[9];
-#U(2.09455838395606,-2.12582868535215,3.27717936377974) q0[2];
-Rz(2.09455838395606) | q0[2];
-Ry(-2.12582868535215) | q0[2];
-Rz(3.27717936377974) | q0[2];
-#U(2.40285550266657,-0.589467805378234,2.35309199445041) q0[3];
-Rz(2.40285550266657) | q0[3];
-Ry(-0.589467805378234) | q0[3];
-Rz(2.35309199445041) | q0[3];
-#U(1.62288493568427,-1.15153565473985,-1.62550203854208) q0[5];
-Rz(1.62288493568427) | q0[5];
-Ry(-1.15153565473985) | q0[5];
-Rz(-1.62550203854208) | q0[5];
-#CX q0[5],q0[3];
-CNOT | (q0[5], q0[3])
-#U(0,0,1.58599353554815) q0[3];
-Rz(1.58599353554815) | q0[3];
-#U(-0.667605386611402,0.0,0.0) q0[5];
-Rz(-0.667605386611402) | q0[5];
-#CX q0[3],q0[5];
-CNOT | (q0[3], q0[5])
-#U(-0.331315773081619,0.0,0.0) q0[5];
-Rz(-0.331315773081619) | q0[5];
-#CX q0[5],q0[3];
-CNOT | (q0[5], q0[3])
-#U(2.74791697666427,1.50178354062613,-1.40604699813349) q0[3];
-Rz(2.74791697666427) | q0[3];
-Ry(1.50178354062613) | q0[3];
-Rz(-1.40604699813349) | q0[3];
-#U(1.50825708893894,-1.15052577811813,-3.59403879349315) q0[5];
-Rz(1.50825708893894) | q0[5];
-Ry(-1.15052577811813) | q0[5];
-Rz(-3.59403879349315) | q0[5];
-#U(2.40210743968880,2.43794117924637,-1.30971653138287) q0[0];
-Rz(2.40210743968880) | q0[0];
-Ry(2.43794117924637) | q0[0];
-Rz(-1.30971653138287) | q0[0];
-#U(1.43449381414803,1.65571668866816,-0.876771793481911) q0[7];
-Rz(1.43449381414803) | q0[7];
-Ry(1.65571668866816) | q0[7];
-Rz(-0.876771793481911) | q0[7];
-#CX q0[7],q0[0];
-CNOT | (q0[7], q0[0])
-#U(0,0,2.99588255581648) q0[0];
-Rz(2.99588255581648) | q0[0];
-#U(-1.16012824307334,0.0,0.0) q0[7];
-Rz(-1.16012824307334) | q0[7];
-#CX q0[0],q0[7];
-CNOT | (q0[0], q0[7])
-#U(2.01651578782364,0.0,0.0) q0[7];
-Rz(2.01651578782364) | q0[7];
-#CX q0[7],q0[0];
-CNOT | (q0[7], q0[0])
-#U(1.81528881604129,1.14898207359622,-2.40928021666194) q0[0];
-Rz(1.81528881604129) | q0[0];
-Ry(1.14898207359622) | q0[0];
-Rz(-2.40928021666194) | q0[0];
-#U(2.54324982685206,1.06802053908671,4.22194925449694) q0[7];
-Rz(2.54324982685206) | q0[7];
-Ry(1.06802053908671) | q0[7];
-Rz(4.22194925449694) | q0[7];
-#U(1.41040073039647,-0.826479181291259,1.67028545759254) q0[3];
-Rz(1.41040073039647) | q0[3];
-Ry(-0.826479181291259) | q0[3];
-Rz(1.67028545759254) | q0[3];
-#U(1.59246725559313,-2.35671500938234,-2.76766064352221) q0[9];
-Rz(1.59246725559313) | q0[9];
-Ry(-2.35671500938234) | q0[9];
-Rz(-2.76766064352221) | q0[9];
-#CX q0[9],q0[3];
-CNOT | (q0[9], q0[3])
-#U(0,0,-0.730564215574114) q0[3];
-Rz(-0.730564215574114) | q0[3];
-#U(0.398901779315231,0.0,0.0) q0[9];
-Rz(0.398901779315231) | q0[9];
-#CX q0[3],q0[9];
-CNOT | (q0[3], q0[9])
-#U(4.23966918620213,0.0,0.0) q0[9];
-Rz(4.23966918620213) | q0[9];
-#CX q0[9],q0[3];
-CNOT | (q0[9], q0[3])
-#U(2.22846832672376,0.698672217538434,-3.28523990081905) q0[3];
-Rz(2.22846832672376) | q0[3];
-Ry(0.698672217538434) | q0[3];
-Rz(-3.28523990081905) | q0[3];
-#U(0.860071788748449,-0.587111047190746,-0.102703593292678) q0[9];
-Rz(0.860071788748449) | q0[9];
-Ry(-0.587111047190746) | q0[9];
-Rz(-0.102703593292678) | q0[9];
-#U(1.84415129254918,-1.71018594230838,-0.592171845444967) q0[2];
-Rz(1.84415129254918) | q0[2];
-Ry(-1.71018594230838) | q0[2];
-Rz(-0.592171845444967) | q0[2];
-#U(1.27035348860092,-4.05237577930272,-0.108484402457318) q0[7];
-Rz(1.27035348860092) | q0[7];
-Ry(-4.05237577930272) | q0[7];
-Rz(-0.108484402457318) | q0[7];
-#CX q0[7],q0[2];
-CNOT | (q0[7], q0[2])
-#U(0,0,1.33796377110171) q0[2];
-Rz(1.33796377110171) | q0[2];
-#U(-0.00954028078021674,0.0,0.0) q0[7];
-Rz(-0.00954028078021674) | q0[7];
-#CX q0[2],q0[7];
-CNOT | (q0[2], q0[7])
-#U(2.45972582015753,0.0,0.0) q0[7];
-Rz(2.45972582015753) | q0[7];
-#CX q0[7],q0[2];
-CNOT | (q0[7], q0[2])
-#U(1.68303728637402,3.82254954086331,-0.712040155102228) q0[2];
-Rz(1.68303728637402) | q0[2];
-Ry(3.82254954086331) | q0[2];
-Rz(-0.712040155102228) | q0[2];
-#U(2.22454747763957,-2.04546509710341,2.15200826107622) q0[7];
-Rz(2.22454747763957) | q0[7];
-Ry(-2.04546509710341) | q0[7];
-Rz(2.15200826107622) | q0[7];
-#U(1.92992972202601,2.98949997521693,-2.31824292571283) q0[4];
-Rz(1.92992972202601) | q0[4];
-Ry(2.98949997521693) | q0[4];
-Rz(-2.31824292571283) | q0[4];
-#U(1.26413998546826,1.53594750550518,-1.40006420180873) q0[1];
-Rz(1.26413998546826) | q0[1];
-Ry(1.53594750550518) | q0[1];
-Rz(-1.40006420180873) | q0[1];
-#CX q0[1],q0[4];
-CNOT | (q0[1], q0[4])
-#U(0,0,0.0422195504457192) q0[4];
-Rz(0.0422195504457192) | q0[4];
-#U(-1.32115643443226,0.0,0.0) q0[1];
-Rz(-1.32115643443226) | q0[1];
-#CX q0[4],q0[1];
-CNOT | (q0[4], q0[1])
-#U(1.94668765058152,0.0,0.0) q0[1];
-Rz(1.94668765058152) | q0[1];
-#CX q0[1],q0[4];
-CNOT | (q0[1], q0[4])
-#U(0.424995336618577,0.995280600570887,-0.734583916673127) q0[4];
-Rz(0.424995336618577) | q0[4];
-Ry(0.995280600570887) | q0[4];
-Rz(-0.734583916673127) | q0[4];
-#U(1.99284209379717,5.00094595768183,1.19762754169712) q0[1];
-Rz(1.99284209379717) | q0[1];
-Ry(5.00094595768183) | q0[1];
-Rz(1.19762754169712) | q0[1];
-#U(2.40486596368620,-2.30947432295757,3.66515118261960) q0[6];
-Rz(2.40486596368620) | q0[6];
-Ry(-2.30947432295757) | q0[6];
-Rz(3.66515118261960) | q0[6];
-#U(0.814793490824071,1.53033774888040,1.29690707853448) q0[0];
-Rz(0.814793490824071) | q0[0];
-Ry(1.53033774888040) | q0[0];
-Rz(1.29690707853448) | q0[0];
-#CX q0[0],q0[6];
-CNOT | (q0[0], q0[6])
-#U(0,0,1.75274489270389) q0[6];
-Rz(1.75274489270389) | q0[6];
-#U(-2.90698526003976,0.0,0.0) q0[0];
-Rz(-2.90698526003976) | q0[0];
-#CX q0[6],q0[0];
-CNOT | (q0[6], q0[0])
-#U(0.711365577259516,0.0,0.0) q0[0];
-Rz(0.711365577259516) | q0[0];
-#CX q0[0],q0[6];
-CNOT | (q0[0], q0[6])
-#U(2.57176694250477,-0.857433758600263,0.478739690564174) q0[6];
-Rz(2.57176694250477) | q0[6];
-Ry(-0.857433758600263) | q0[6];
-Rz(0.478739690564174) | q0[6];
-#U(1.63969995677569,2.35656091397537,0.127556463720104) q0[0];
-Rz(1.63969995677569) | q0[0];
-Ry(2.35656091397537) | q0[0];
-Rz(0.127556463720104) | q0[0];
-#U(2.01910798435696,2.08518383996055,-2.81879746352554) q0[8];
-Rz(2.01910798435696) | q0[8];
-Ry(2.08518383996055) | q0[8];
-Rz(-2.81879746352554) | q0[8];
-#U(1.86416902840682,-2.64934205601315,2.70092079639789) q0[5];
-Rz(1.86416902840682) | q0[5];
-Ry(-2.64934205601315) | q0[5];
-Rz(2.70092079639789) | q0[5];
-#CX q0[5],q0[8];
-CNOT | (q0[5], q0[8])
-#U(0,0,-0.0756513967830656) q0[8];
-Rz(-0.0756513967830656) | q0[8];
-#U(-2.47532365997556,0.0,0.0) q0[5];
-Rz(-2.47532365997556) | q0[5];
-#CX q0[8],q0[5];
-CNOT | (q0[8], q0[5])
-#U(1.47967712411787,0.0,0.0) q0[5];
-Rz(1.47967712411787) | q0[5];
-#CX q0[5],q0[8];
-CNOT | (q0[5], q0[8])
-#U(2.17566049955499,0.518010878474775,-3.00742758422112) q0[8];
-Rz(2.17566049955499) | q0[8];
-Ry(0.518010878474775) | q0[8];
-Rz(-3.00742758422112) | q0[8];
-#U(1.73177742547724,-0.857318244701762,3.67356071602780) q0[5];
-Rz(1.73177742547724) | q0[5];
-Ry(-0.857318244701762) | q0[5];
-Rz(3.67356071602780) | q0[5];
-#U(1.13156061898750,2.27718116856779,-1.80148256259717) q0[6];
-Rz(1.13156061898750) | q0[6];
-Ry(2.27718116856779) | q0[6];
-Rz(-1.80148256259717) | q0[6];
-#U(0.317870413085444,0.474428289749719,-1.57237860807379) q0[4];
-Rz(0.317870413085444) | q0[4];
-Ry(0.474428289749719) | q0[4];
-Rz(-1.57237860807379) | q0[4];
-#CX q0[4],q0[6];
-CNOT | (q0[4], q0[6])
-#U(0,0,0.725496754302604) q0[6];
-Rz(0.725496754302604) | q0[6];
-#U(-3.16956953580261,0.0,0.0) q0[4];
-Rz(-3.16956953580261) | q0[4];
-#CX q0[6],q0[4];
-CNOT | (q0[6], q0[4])
-#U(1.54963245867998,0.0,0.0) q0[4];
-Rz(1.54963245867998) | q0[4];
-#CX q0[4],q0[6];
-CNOT | (q0[4], q0[6])
-#U(1.15724522850173,2.73199994429998,-1.44119724804122) q0[6];
-Rz(1.15724522850173) | q0[6];
-Ry(2.73199994429998) | q0[6];
-Rz(-1.44119724804122) | q0[6];
-#U(0.947563652474974,-1.03537396858126,-3.73404002659393) q0[4];
-Rz(0.947563652474974) | q0[4];
-Ry(-1.03537396858126) | q0[4];
-Rz(-3.73404002659393) | q0[4];
-#U(1.97721338473687,-1.80962232406986,-0.100011003952981) q0[9];
-Rz(1.97721338473687) | q0[9];
-Ry(-1.80962232406986) | q0[9];
-Rz(-0.100011003952981) | q0[9];
-#U(1.20098331325541,-2.09884045602267,-1.19784134795245) q0[7];
-Rz(1.20098331325541) | q0[7];
-Ry(-2.09884045602267) | q0[7];
-Rz(-1.19784134795245) | q0[7];
-#CX q0[7],q0[9];
-CNOT | (q0[7], q0[9])
-#U(0,0,0.840063937535285) q0[9];
-Rz(0.840063937535285) | q0[9];
-#U(-0.0996180377594906,0.0,0.0) q0[7];
-Rz(-0.0996180377594906) | q0[7];
-#CX q0[9],q0[7];
-CNOT | (q0[9], q0[7])
-#U(2.57448215072050,0.0,0.0) q0[7];
-Rz(2.57448215072050) | q0[7];
-#CX q0[7],q0[9];
-CNOT | (q0[7], q0[9])
-#U(1.23276200199611,2.99630079227189,-1.96243581352837) q0[9];
-Rz(1.23276200199611) | q0[9];
-Ry(2.99630079227189) | q0[9];
-Rz(-1.96243581352837) | q0[9];
-#U(1.25456518647770,-2.09502051241465,-1.43722621387689) q0[7];
-Rz(1.25456518647770) | q0[7];
-Ry(-2.09502051241465) | q0[7];
-Rz(-1.43722621387689) | q0[7];
-#U(1.07885759605728,1.42849299806781,-1.00383778893748) q0[3];
-Rz(1.07885759605728) | q0[3];
-Ry(1.42849299806781) | q0[3];
-Rz(-1.00383778893748) | q0[3];
-#U(2.13806349166564,-1.30597023321018,-3.66782323223535) q0[1];
-Rz(2.13806349166564) | q0[1];
-Ry(-1.30597023321018) | q0[1];
-Rz(-3.66782323223535) | q0[1];
-#CX q0[1],q0[3];
-CNOT | (q0[1], q0[3])
-#U(0,0,1.54043406070446) q0[3];
-Rz(1.54043406070446) | q0[3];
-#U(-3.46057779272240,0.0,0.0) q0[1];
-Rz(-3.46057779272240) | q0[1];
-#CX q0[3],q0[1];
-CNOT | (q0[3], q0[1])
-#U(2.15975170961543,0.0,0.0) q0[1];
-Rz(2.15975170961543) | q0[1];
-#CX q0[1],q0[3];
-CNOT | (q0[1], q0[3])
-#U(2.43646598650902,1.32361646551313,-4.21645773181308) q0[3];
-Rz(2.43646598650902) | q0[3];
-Ry(1.32361646551313) | q0[3];
-Rz(-4.21645773181308) | q0[3];
-#U(2.49431638418182,-1.58879925654948,0.657203428818291) q0[1];
-Rz(2.49431638418182) | q0[1];
-Ry(-1.58879925654948) | q0[1];
-Rz(0.657203428818291) | q0[1];
-#U(1.06355259050574,-1.28136280546724,0.173510727853247) q0[5];
-Rz(1.06355259050574) | q0[5];
-Ry(-1.28136280546724) | q0[5];
-Rz(0.173510727853247) | q0[5];
-#U(1.47179619385959,-1.86974770184395,1.38554586141494) q0[8];
-Rz(1.47179619385959) | q0[8];
-Ry(-1.86974770184395) | q0[8];
-Rz(1.38554586141494) | q0[8];
-#CX q0[8],q0[5];
-CNOT | (q0[8], q0[5])
-#U(0,0,3.12245077579140) q0[5];
-Rz(3.12245077579140) | q0[5];
-#U(-3.80625284838098,0.0,0.0) q0[8];
-Rz(-3.80625284838098) | q0[8];
-#CX q0[5],q0[8];
-CNOT | (q0[5], q0[8])
-#U(-0.815896014740553,0.0,0.0) q0[8];
-Rz(-0.815896014740553) | q0[8];
-#CX q0[8],q0[5];
-CNOT | (q0[8], q0[5])
-#U(2.26641077401172,-3.90004372011091,2.22559990044196) q0[5];
-Rz(2.26641077401172) | q0[5];
-Ry(-3.90004372011091) | q0[5];
-Rz(2.22559990044196) | q0[5];
-#U(0.974616843163170,-4.01854244510685,1.38289165509869) q0[8];
-Rz(0.974616843163170) | q0[8];
-Ry(-4.01854244510685) | q0[8];
-Rz(1.38289165509869) | q0[8];
-#U(1.31860987426997,-0.380763821019618,0.917225998165995) q0[0];
-Rz(1.31860987426997) | q0[0];
-Ry(-0.380763821019618) | q0[0];
-Rz(0.917225998165995) | q0[0];
-#U(1.81276577706631,-3.28433142531238,-0.372582622006046) q0[2];
-Rz(1.81276577706631) | q0[2];
-Ry(-3.28433142531238) | q0[2];
-Rz(-0.372582622006046) | q0[2];
-#CX q0[2],q0[0];
-CNOT | (q0[2], q0[0])
-#U(0,0,4.29776897778576) q0[0];
-Rz(4.29776897778576) | q0[0];
-#U(-3.69350771219233,0.0,0.0) q0[2];
-Rz(-3.69350771219233) | q0[2];
-#CX q0[0],q0[2];
-CNOT | (q0[0], q0[2])
-#U(-0.455751956699722,0.0,0.0) q0[2];
-Rz(-0.455751956699722) | q0[2];
-#CX q0[2],q0[0];
-CNOT | (q0[2], q0[0])
-#U(1.87864821752009,-0.831074681994016,1.86145389124768) q0[0];
-Rz(1.87864821752009) | q0[0];
-Ry(-0.831074681994016) | q0[0];
-Rz(1.86145389124768) | q0[0];
-#U(1.50626637484432,-0.646234152746133,-0.894267593728222) q0[2];
-Rz(1.50626637484432) | q0[2];
-Ry(-0.646234152746133) | q0[2];
-Rz(-0.894267593728222) | q0[2];
-#U(1.99671654098506,3.55409320471735,-0.706608439925495) q0[1];
-Rz(1.99671654098506) | q0[1];
-Ry(3.55409320471735) | q0[1];
-Rz(-0.706608439925495) | q0[1];
-#U(2.03674977349260,2.47644035450152,-1.46013766541673) q0[3];
-Rz(2.03674977349260) | q0[3];
-Ry(2.47644035450152) | q0[3];
-Rz(-1.46013766541673) | q0[3];
-#CX q0[3],q0[1];
-CNOT | (q0[3], q0[1])
-#U(0,0,1.74599378281325) q0[1];
-Rz(1.74599378281325) | q0[1];
-#U(-2.68490271922246,0.0,0.0) q0[3];
-Rz(-2.68490271922246) | q0[3];
-#CX q0[1],q0[3];
-CNOT | (q0[1], q0[3])
-#U(1.10558345205590,0.0,0.0) q0[3];
-Rz(1.10558345205590) | q0[3];
-#CX q0[3],q0[1];
-CNOT | (q0[3], q0[1])
-#U(2.78075130019223,-0.189120391961622,-2.26508642057054) q0[1];
-Rz(2.78075130019223) | q0[1];
-Ry(-0.189120391961622) | q0[1];
-Rz(-2.26508642057054) | q0[1];
-#U(1.60903859800909,-0.700381547740548,4.75132082956504) q0[3];
-Rz(1.60903859800909) | q0[3];
-Ry(-0.700381547740548) | q0[3];
-Rz(4.75132082956504) | q0[3];
-#U(1.61193778027899,1.65911944138257,-2.39010906203369) q0[4];
-Rz(1.61193778027899) | q0[4];
-Ry(1.65911944138257) | q0[4];
-Rz(-2.39010906203369) | q0[4];
-#U(2.21028250152187,2.10780193245699,-3.55973129269899) q0[2];
-Rz(2.21028250152187) | q0[2];
-Ry(2.10780193245699) | q0[2];
-Rz(-3.55973129269899) | q0[2];
-#CX q0[2],q0[4];
-CNOT | (q0[2], q0[4])
-#U(0,0,3.98331450438229) q0[4];
-Rz(3.98331450438229) | q0[4];
-#U(-3.02842255850596,0.0,0.0) q0[2];
-Rz(-3.02842255850596) | q0[2];
-#CX q0[4],q0[2];
-CNOT | (q0[4], q0[2])
-#U(-0.575561507376651,0.0,0.0) q0[2];
-Rz(-0.575561507376651) | q0[2];
-#CX q0[2],q0[4];
-CNOT | (q0[2], q0[4])
-#U(1.57790018338351,1.04905288426232,1.10723176731755) q0[4];
-Rz(1.57790018338351) | q0[4];
-Ry(1.04905288426232) | q0[4];
-Rz(1.10723176731755) | q0[4];
-#U(2.11403763375287,-2.16508411636333,0.717621408364864) q0[2];
-Rz(2.11403763375287) | q0[2];
-Ry(-2.16508411636333) | q0[2];
-Rz(0.717621408364864) | q0[2];
-#U(1.93143171156133,-3.18343744157771,1.91494300700292) q0[0];
-Rz(1.93143171156133) | q0[0];
-Ry(-3.18343744157771) | q0[0];
-Rz(1.91494300700292) | q0[0];
-#U(0.741965139721785,3.18976020804651,-1.00405761189597) q0[6];
-Rz(0.741965139721785) | q0[6];
-Ry(3.18976020804651) | q0[6];
-Rz(-1.00405761189597) | q0[6];
-#CX q0[6],q0[0];
-CNOT | (q0[6], q0[0])
-#U(0,0,0.426026077835793) q0[0];
-Rz(0.426026077835793) | q0[0];
-#U(-0.300990473340881,0.0,0.0) q0[6];
-Rz(-0.300990473340881) | q0[6];
-#CX q0[0],q0[6];
-CNOT | (q0[0], q0[6])
-#U(1.36854780929068,0.0,0.0) q0[6];
-Rz(1.36854780929068) | q0[6];
-#CX q0[6],q0[0];
-CNOT | (q0[6], q0[0])
-#U(1.13854623507452,3.04952325258543,-1.32415229953516) q0[0];
-Rz(1.13854623507452) | q0[0];
-Ry(3.04952325258543) | q0[0];
-Rz(-1.32415229953516) | q0[0];
-#U(2.39103814923015,1.77508122044363,-0.689142110951366) q0[6];
-Rz(2.39103814923015) | q0[6];
-Ry(1.77508122044363) | q0[6];
-Rz(-0.689142110951366) | q0[6];
-#U(1.66193347175956,0.618203912431520,1.08198675324451) q0[7];
-Rz(1.66193347175956) | q0[7];
-Ry(0.618203912431520) | q0[7];
-Rz(1.08198675324451) | q0[7];
-#U(0.0842008818928847,-4.54877109113564,-0.219014209894349) q0[9];
-Rz(0.0842008818928847) | q0[9];
-Ry(-4.54877109113564) | q0[9];
-Rz(-0.219014209894349) | q0[9];
-#CX q0[9],q0[7];
-CNOT | (q0[9], q0[7])
-#U(0,0,1.99895848962412) q0[7];
-Rz(1.99895848962412) | q0[7];
-#U(-2.51802379310628,0.0,0.0) q0[9];
-Rz(-2.51802379310628) | q0[9];
-#CX q0[7],q0[9];
-CNOT | (q0[7], q0[9])
-#U(0.0782296147918011,0.0,0.0) q0[9];
-Rz(0.0782296147918011) | q0[9];
-#CX q0[9],q0[7];
-CNOT | (q0[9], q0[7])
-#U(2.54848418955689,2.22912691289522,-2.09152286271362) q0[7];
-Rz(2.54848418955689) | q0[7];
-Ry(2.22912691289522) | q0[7];
-Rz(-2.09152286271362) | q0[7];
-#U(1.72635388443217,-0.230113950515224,-2.50530579823871) q0[9];
-Rz(1.72635388443217) | q0[9];
-Ry(-0.230113950515224) | q0[9];
-Rz(-2.50530579823871) | q0[9];
-#U(2.35592181955447,2.64139336948466,-1.70912931677511) q0[5];
-Rz(2.35592181955447) | q0[5];
-Ry(2.64139336948466) | q0[5];
-Rz(-1.70912931677511) | q0[5];
-#U(1.45497045024009,2.12554365244366,-2.68948251609080) q0[8];
-Rz(1.45497045024009) | q0[8];
-Ry(2.12554365244366) | q0[8];
-Rz(-2.68948251609080) | q0[8];
-#CX q0[8],q0[5];
-CNOT | (q0[8], q0[5])
-#U(0,0,1.14488671212708) q0[5];
-Rz(1.14488671212708) | q0[5];
-#U(-3.41331362457461,0.0,0.0) q0[8];
-Rz(-3.41331362457461) | q0[8];
-#CX q0[5],q0[8];
-CNOT | (q0[5], q0[8])
-#U(1.54880829931241,0.0,0.0) q0[8];
-Rz(1.54880829931241) | q0[8];
-#CX q0[8],q0[5];
-CNOT | (q0[8], q0[5])
-#U(1.32117778157928,1.00384763656143,-2.27412904527513) q0[5];
-Rz(1.32117778157928) | q0[5];
-Ry(1.00384763656143) | q0[5];
-Rz(-2.27412904527513) | q0[5];
-#U(1.31998998939091,2.43119846813663,3.43528497282949) q0[8];
-Rz(1.31998998939091) | q0[8];
-Ry(2.43119846813663) | q0[8];
-Rz(3.43528497282949) | q0[8];
-#U(2.10413467122134,1.76459940726736,-2.42591559938147) q0[3];
-Rz(2.10413467122134) | q0[3];
-Ry(1.76459940726736) | q0[3];
-Rz(-2.42591559938147) | q0[3];
-#U(1.49665076822509,-2.89722158530940,2.80839446653340) q0[2];
-Rz(1.49665076822509) | q0[2];
-Ry(-2.89722158530940) | q0[2];
-Rz(2.80839446653340) | q0[2];
-#CX q0[2],q0[3];
-CNOT | (q0[2], q0[3])
-#U(0,0,1.38500693133039) q0[3];
-Rz(1.38500693133039) | q0[3];
-#U(-3.24161893566252,0.0,0.0) q0[2];
-Rz(-3.24161893566252) | q0[2];
-#CX q0[3],q0[2];
-CNOT | (q0[3], q0[2])
-#U(2.70288610276879,0.0,0.0) q0[2];
-Rz(2.70288610276879) | q0[2];
-#CX q0[2],q0[3];
-CNOT | (q0[2], q0[3])
-#U(1.33374452315766,-1.78804555860672,1.97282408974588) q0[3];
-Rz(1.33374452315766) | q0[3];
-Ry(-1.78804555860672) | q0[3];
-Rz(1.97282408974588) | q0[3];
-#U(1.68230580344247,0.238023722571170,5.85167277605247) q0[2];
-Rz(1.68230580344247) | q0[2];
-Ry(0.238023722571170) | q0[2];
-Rz(5.85167277605247) | q0[2];
-#U(2.14660458533472,1.20579279599353,-2.32907400317017) q0[8];
-Rz(2.14660458533472) | q0[8];
-Ry(1.20579279599353) | q0[8];
-Rz(-2.32907400317017) | q0[8];
-#U(2.04168821115085,2.12582897382556,-3.21248019668648) q0[6];
-Rz(2.04168821115085) | q0[6];
-Ry(2.12582897382556) | q0[6];
-Rz(-3.21248019668648) | q0[6];
-#CX q0[6],q0[8];
-CNOT | (q0[6], q0[8])
-#U(0,0,3.50748655826933) q0[8];
-Rz(3.50748655826933) | q0[8];
-#U(-1.53058348994348,0.0,0.0) q0[6];
-Rz(-1.53058348994348) | q0[6];
-#CX q0[8],q0[6];
-CNOT | (q0[8], q0[6])
-#U(2.48340606746683,0.0,0.0) q0[6];
-Rz(2.48340606746683) | q0[6];
-#CX q0[6],q0[8];
-CNOT | (q0[6], q0[8])
-#U(2.15690763550813,-0.867131716479593,3.35745057667818) q0[8];
-Rz(2.15690763550813) | q0[8];
-Ry(-0.867131716479593) | q0[8];
-Rz(3.35745057667818) | q0[8];
-#U(1.62476116334594,-3.65040270190010,-1.09524023900228) q0[6];
-Rz(1.62476116334594) | q0[6];
-Ry(-3.65040270190010) | q0[6];
-Rz(-1.09524023900228) | q0[6];
-#U(1.08133971636642,-2.84007443986253,0.283971143010596) q0[7];
-Rz(1.08133971636642) | q0[7];
-Ry(-2.84007443986253) | q0[7];
-Rz(0.283971143010596) | q0[7];
-#U(1.98743883471224,-3.15586429954020,-0.653915823829388) q0[9];
-Rz(1.98743883471224) | q0[9];
-Ry(-3.15586429954020) | q0[9];
-Rz(-0.653915823829388) | q0[9];
-#CX q0[9],q0[7];
-CNOT | (q0[9], q0[7])
-#U(0,0,2.13330310054202) q0[7];
-Rz(2.13330310054202) | q0[7];
-#U(0.104677731471594,0.0,0.0) q0[9];
-Rz(0.104677731471594) | q0[9];
-#CX q0[7],q0[9];
-CNOT | (q0[7], q0[9])
-#U(1.49472560430759,0.0,0.0) q0[9];
-Rz(1.49472560430759) | q0[9];
-#CX q0[9],q0[7];
-CNOT | (q0[9], q0[7])
-#U(0.476428791228162,-3.21716666999406,1.97529028174835) q0[7];
-Rz(0.476428791228162) | q0[7];
-Ry(-3.21716666999406) | q0[7];
-Rz(1.97529028174835) | q0[7];
-#U(1.60515799138471,-3.03709059586943,2.59766912590832) q0[9];
-Rz(1.60515799138471) | q0[9];
-Ry(-3.03709059586943) | q0[9];
-Rz(2.59766912590832) | q0[9];
-#U(0.792203625906823,0.761639607717834,-0.348685771378711) q0[1];
-Rz(0.792203625906823) | q0[1];
-Ry(0.761639607717834) | q0[1];
-Rz(-0.348685771378711) | q0[1];
-#U(1.05378472144179,-0.238171338443441,-0.896810320008973) q0[4];
-Rz(1.05378472144179) | q0[4];
-Ry(-0.238171338443441) | q0[4];
-Rz(-0.896810320008973) | q0[4];
-#CX q0[4],q0[1];
-CNOT | (q0[4], q0[1])
-#U(0,0,1.61359931707247) q0[1];
-Rz(1.61359931707247) | q0[1];
-#U(-2.33364828827733,0.0,0.0) q0[4];
-Rz(-2.33364828827733) | q0[4];
-#CX q0[1],q0[4];
-CNOT | (q0[1], q0[4])
-#U(3.02671402063799,0.0,0.0) q0[4];
-Rz(3.02671402063799) | q0[4];
-#CX q0[4],q0[1];
-CNOT | (q0[4], q0[1])
-#U(1.67668599529984,3.90190029809733,-1.94222833210093) q0[1];
-Rz(1.67668599529984) | q0[1];
-Ry(3.90190029809733) | q0[1];
-Rz(-1.94222833210093) | q0[1];
-#U(1.19803802000337,0.109413096624148,4.53189106844810) q0[4];
-Rz(1.19803802000337) | q0[4];
-Ry(0.109413096624148) | q0[4];
-Rz(4.53189106844810) | q0[4];
-#U(1.17622150577850,0.710441000035738,-2.53716165885516) q0[0];
-Rz(1.17622150577850) | q0[0];
-Ry(0.710441000035738) | q0[0];
-Rz(-2.53716165885516) | q0[0];
-#U(1.76068312643064,3.15779055057136,-2.40563401301631) q0[5];
-Rz(1.76068312643064) | q0[5];
-Ry(3.15779055057136) | q0[5];
-Rz(-2.40563401301631) | q0[5];
-#CX q0[5],q0[0];
-CNOT | (q0[5], q0[0])
-#U(0,0,3.76313783628596) q0[0];
-Rz(3.76313783628596) | q0[0];
-#U(-1.04115462438174,0.0,0.0) q0[5];
-Rz(-1.04115462438174) | q0[5];
-#CX q0[0],q0[5];
-CNOT | (q0[0], q0[5])
-#U(1.94456033360708,0.0,0.0) q0[5];
-Rz(1.94456033360708) | q0[5];
-#CX q0[5],q0[0];
-CNOT | (q0[5], q0[0])
-#U(1.85598765913660,-0.428758344637324,3.55645944714666) q0[0];
-Rz(1.85598765913660) | q0[0];
-Ry(-0.428758344637324) | q0[0];
-Rz(3.55645944714666) | q0[0];
-#U(1.48874740714208,0.506223587544506,3.58694976804594) q0[5];
-Rz(1.48874740714208) | q0[5];
-Ry(0.506223587544506) | q0[5];
-Rz(3.58694976804594) | q0[5];
-#U(0.862674925696716,-0.890852679087657,-0.790815979132237) q0[6];
-Rz(0.862674925696716) | q0[6];
-Ry(-0.890852679087657) | q0[6];
-Rz(-0.790815979132237) | q0[6];
-#U(1.55810757997084,0.980829128657701,-5.10845468498816) q0[0];
-Rz(1.55810757997084) | q0[0];
-Ry(0.980829128657701) | q0[0];
-Rz(-5.10845468498816) | q0[0];
-#CX q0[0],q0[6];
-CNOT | (q0[0], q0[6])
-#U(0,0,2.80183395235846) q0[6];
-Rz(2.80183395235846) | q0[6];
-#U(-1.58682034003030,0.0,0.0) q0[0];
-Rz(-1.58682034003030) | q0[0];
-#CX q0[6],q0[0];
-CNOT | (q0[6], q0[0])
-#U(0.473094119026529,0.0,0.0) q0[0];
-Rz(0.473094119026529) | q0[0];
-#CX q0[0],q0[6];
-CNOT | (q0[0], q0[6])
-#U(2.00355972827992,3.76383057000117,-1.20056270436489) q0[6];
-Rz(2.00355972827992) | q0[6];
-Ry(3.76383057000117) | q0[6];
-Rz(-1.20056270436489) | q0[6];
-#U(0.777346187005935,-4.87649395309460,-0.668784149660834) q0[0];
-Rz(0.777346187005935) | q0[0];
-Ry(-4.87649395309460) | q0[0];
-Rz(-0.668784149660834) | q0[0];
-#U(1.64898912836646,0.306924261354068,2.08772249118987) q0[8];
-Rz(1.64898912836646) | q0[8];
-Ry(0.306924261354068) | q0[8];
-Rz(2.08772249118987) | q0[8];
-#U(1.77285757348099,-2.01019595272478,-1.30869274130165) q0[3];
-Rz(1.77285757348099) | q0[3];
-Ry(-2.01019595272478) | q0[3];
-Rz(-1.30869274130165) | q0[3];
-#CX q0[3],q0[8];
-CNOT | (q0[3], q0[8])
-#U(0,0,3.84448692727029) q0[8];
-Rz(3.84448692727029) | q0[8];
-#U(-1.38689944136488,0.0,0.0) q0[3];
-Rz(-1.38689944136488) | q0[3];
-#CX q0[8],q0[3];
-CNOT | (q0[8], q0[3])
-#U(2.21522345321645,0.0,0.0) q0[3];
-Rz(2.21522345321645) | q0[3];
-#CX q0[3],q0[8];
-CNOT | (q0[3], q0[8])
-#U(2.71172956943155,3.48856981075694,-1.12523423753372) q0[8];
-Rz(2.71172956943155) | q0[8];
-Ry(3.48856981075694) | q0[8];
-Rz(-1.12523423753372) | q0[8];
-#U(1.30959705199769,-2.03681311744479,-2.06196932864980) q0[3];
-Rz(1.30959705199769) | q0[3];
-Ry(-2.03681311744479) | q0[3];
-Rz(-2.06196932864980) | q0[3];
-#U(1.82139069393684,2.00576505756883,-2.55459717544893) q0[5];
-Rz(1.82139069393684) | q0[5];
-Ry(2.00576505756883) | q0[5];
-Rz(-2.55459717544893) | q0[5];
-#U(0.653653144007384,2.60878654389954,-2.15002636625922) q0[7];
-Rz(0.653653144007384) | q0[7];
-Ry(2.60878654389954) | q0[7];
-Rz(-2.15002636625922) | q0[7];
-#CX q0[7],q0[5];
-CNOT | (q0[7], q0[5])
-#U(0,0,1.60617528274604) q0[5];
-Rz(1.60617528274604) | q0[5];
-#U(-2.84909800945104,0.0,0.0) q0[7];
-Rz(-2.84909800945104) | q0[7];
-#CX q0[5],q0[7];
-CNOT | (q0[5], q0[7])
-#U(0.547520293700537,0.0,0.0) q0[7];
-Rz(0.547520293700537) | q0[7];
-#CX q0[7],q0[5];
-CNOT | (q0[7], q0[5])
-#U(2.34437645139780,0.785937938310707,-2.22680057564874) q0[5];
-Rz(2.34437645139780) | q0[5];
-Ry(0.785937938310707) | q0[5];
-Rz(-2.22680057564874) | q0[5];
-#U(2.73036323176218,-0.229366111563844,-1.54514738940961) q0[7];
-Rz(2.73036323176218) | q0[7];
-Ry(-0.229366111563844) | q0[7];
-Rz(-1.54514738940961) | q0[7];
-#U(1.45757646904717,0.0269157815567683,-1.17830022153592) q0[1];
-Rz(1.45757646904717) | q0[1];
-Ry(0.0269157815567683) | q0[1];
-Rz(-1.17830022153592) | q0[1];
-#U(1.37712786145726,-2.87129954997340,0.549277939699940) q0[4];
-Rz(1.37712786145726) | q0[4];
-Ry(-2.87129954997340) | q0[4];
-Rz(0.549277939699940) | q0[4];
-#CX q0[4],q0[1];
-CNOT | (q0[4], q0[1])
-#U(0,0,2.35766919989753) q0[1];
-Rz(2.35766919989753) | q0[1];
-#U(-1.65827706408371,0.0,0.0) q0[4];
-Rz(-1.65827706408371) | q0[4];
-#CX q0[1],q0[4];
-CNOT | (q0[1], q0[4])
-#U(3.49616190467947,0.0,0.0) q0[4];
-Rz(3.49616190467947) | q0[4];
-#CX q0[4],q0[1];
-CNOT | (q0[4], q0[1])
-#U(1.41368696654425,1.63536645495981,-2.02144154402379) q0[1];
-Rz(1.41368696654425) | q0[1];
-Ry(1.63536645495981) | q0[1];
-Rz(-2.02144154402379) | q0[1];
-#U(1.36477769099758,2.58397271896477,3.27255525135136) q0[4];
-Rz(1.36477769099758) | q0[4];
-Ry(2.58397271896477) | q0[4];
-Rz(3.27255525135136) | q0[4];
-#U(1.44803300875930,-1.16596796924142,-1.03212643976397) q0[9];
-Rz(1.44803300875930) | q0[9];
-Ry(-1.16596796924142) | q0[9];
-Rz(-1.03212643976397) | q0[9];
-#U(1.93132766776404,-3.29458276010450,-0.198135868428953) q0[2];
-Rz(1.93132766776404) | q0[2];
-Ry(-3.29458276010450) | q0[2];
-Rz(-0.198135868428953) | q0[2];
-#CX q0[2],q0[9];
-CNOT | (q0[2], q0[9])
-#U(0,0,1.71557451635185) q0[9];
-Rz(1.71557451635185) | q0[9];
-#U(-2.55797924450771,0.0,0.0) q0[2];
-Rz(-2.55797924450771) | q0[2];
-#CX q0[9],q0[2];
-CNOT | (q0[9], q0[2])
-#U(0.982266965835827,0.0,0.0) q0[2];
-Rz(0.982266965835827) | q0[2];
-#CX q0[2],q0[9];
-CNOT | (q0[2], q0[9])
-#U(1.90961568045603,-1.45211863615260,0.363685360958541) q0[9];
-Rz(1.90961568045603) | q0[9];
-Ry(-1.45211863615260) | q0[9];
-Rz(0.363685360958541) | q0[9];
-#U(1.12230398914785,4.63487787305155,-1.61588893610529) q0[2];
-Rz(1.12230398914785) | q0[2];
-Ry(4.63487787305155) | q0[2];
-Rz(-1.61588893610529) | q0[2];
-#barrier q0[0],q0[1],q0[2],q0[3],q0[4],q0[5],q0[6],q0[7],q0[8],q0[9];
-#measure q0[0] -> c0[0];
-Measure | q0[0]
-c0[0] = int( q0[0] )
-#measure q0[1] -> c0[1];
-Measure | q0[1]
-c0[1] = int( q0[1] )
-#measure q0[2] -> c0[2];
-Measure | q0[2]
-c0[2] = int( q0[2] )
-#measure q0[3] -> c0[3];
-Measure | q0[3]
-c0[3] = int( q0[3] )
-#measure q0[4] -> c0[4];
-Measure | q0[4]
-c0[4] = int( q0[4] )
-#measure q0[5] -> c0[5];
-Measure | q0[5]
-c0[5] = int( q0[5] )
-#measure q0[6] -> c0[6];
-Measure | q0[6]
-c0[6] = int( q0[6] )
-#measure q0[7] -> c0[7];
-Measure | q0[7]
-c0[7] = int( q0[7] )
-#measure q0[8] -> c0[8];
-Measure | q0[8]
-c0[8] = int( q0[8] )
-#measure q0[9] -> c0[9];
-Measure | q0[9]
-c0[9] = int( q0[9] )
-eng.flush()
-elapsed_time = time.time() - start
-print(format(elapsed_time))
