+++ benchmarks/qasm/quantum_volume/ref/quantum_volume_n10_d10_0.qasm.local_qasm_simulator.ref
@@ -1,1 +1,1 @@
-{"1000101110": 1}
+{"1000011011": 1}
